[Chore] Factor out steps into separate module
Problem: wizard_structure contains too many definitions, which
became hard to maintain.

Solution: Put all common steps-related code under `steps` module.
