+++ baking/src/tezos_baking/tezos_voting_wizard.py
@@ -14,6 +14,7 @@ import re
 
 from tezos_baking.wizard_structure import *
 from tezos_baking.util import *
+from tezos_baking.steps import *
 from tezos_baking.validators import Validator
 import tezos_baking.validators as validators
